requirements.txt
Flask==3.1.0
pywebview==6.1
requests==2.32.3
zai-sdk==0.2.1
sniffio==1.3.0
nuitka